requirements.txt
streamlit
pandas
python-dotenv
cohere
numpy
scikit-learn 